+++ Astrocolibri.cs
@@ -57,6 +57,8 @@ namespace ChristophNieswand.NINA.Astrocolibri {
             RaisePropertyChanged(nameof(AstroColibriOptions.JSONFilePath));
             RaisePropertyChanged(nameof(AstroColibriOptions.KeepFilesDays));
             RaisePropertyChanged(nameof(AstroColibriOptions.MinAltitude));
+            RaisePropertyChanged(nameof(AstroColibriOptions.TestMode));
+            RaisePropertyChanged(nameof(AstroColibriOptions.DsoTemplate));
         }
 
         public event PropertyChangedEventHandler PropertyChanged;